move assessment list to own file
